revoke after days metadata accepts float values

+++ impresso/models/specialMembershipDataset.py
@@ -12,7 +12,7 @@ class Metadata(TypedDict, total=False):
 
     modality: Optional[str]  # e.g., "cc_reviewer", "notify_reviewer"
     enableTemporaryAutomaticAcceptance: Optional[bool]
-    revokeAfterDays: Optional[int]
+    revokeAfterDays: Optional[float]
 
 
 class SpecialMembershipDataset(models.Model):
